requirement.txt
pysha3==1.0.2
nem-python==0.1.0
nem-ed25519==0.0.6